init templates
